refactor: improve perf with sqlite
